requirements.txt
Flask
tensorflow
pillow
numpy
gunicorn
pip freeze > requirements.txt
